[Update] auto register model preparer to Service provider boot

+++ src/Commands/MakePreparer.php
@@ -9,7 +9,6 @@
 
 namespace Saad\QueryParser\Commands;
 
-use Illuminate\Console\Command;
 use Saad\Fractal\Commands\BaseMakeCommand;
 
 class MakePreparer extends BaseMakeCommand
@@ -35,20 +34,20 @@ class MakePreparer extends BaseMakeCommand
      */
     public function handle()
     {
-        parent::handle();
+        if (! parent::handle()) {
+            return;
+        }
 
-    	try {
-	    	$this->create("Preparer");
+        try {
+            $this->create("Preparer");
+             $this->addToServiceProviders();
+        } catch (\Exception $exception) {
+            $this->error($exception->getMessage());
+            $this->info('Rolling Back');
 
-            // $this->addToServiceProviders();
-
-	    } catch (\Exception $exception) {
-	    	$this->error($exception->getMessage());
-	    	$this->info('Rolling Back');
-
-	    	$this->delete("Preparer");
-            // $this->removeFromServiceProviders();
-	    }
+            $this->delete("Preparer");
+             $this->removeFromServiceProviders();
+        }
     }
 
     /**
@@ -56,7 +55,7 @@ class MakePreparer extends BaseMakeCommand
      * @return string stubs path
      */
     protected function getStubsPath() {
-    	return __DIR__ . "/../../resources/stubs";
+        return __DIR__ . "/../../resources/stubs";
     }
 
     /**
@@ -64,7 +63,7 @@ class MakePreparer extends BaseMakeCommand
      * @return string Directory name
      */
     protected function getOutputDirectoryName() {
-    	return 'ModelPreparers';
+        return 'ModelPreparers';
     }
 
     /**
@@ -75,7 +74,7 @@ class MakePreparer extends BaseMakeCommand
         $stub_content = $this->filesystem->get($this->getStubsPath() . '/ServiceProvider.stub');
         $stub_content = $this->processStubContent($stub_content);
 
-        $this->updateProviderBinding('/(register\(\).*?\{)/s', "$1 {$stub_content}");
+        $this->updateProviderBinding('/(boot\(\).*?\{)/s', "$1 {$stub_content}");
     }
 
     /**
@@ -92,22 +91,25 @@ class MakePreparer extends BaseMakeCommand
      * @return string pattern
      */
     protected function getBindPattern() {
-        return "/#\sBind\s".$this->model.".*?#\s".$this->model."\sEnd/s";
+        $skipped_full_model = str_replace('\\', '\\\\', $this->full_model);
+        return "/#\sBind\s".$skipped_full_model."\sPreparer.*?#\s".$skipped_full_model."\sPreparer\sEnd/smu";
     }
 
     /**
      * Add or Remove binding to Service Provider
-     * 
-     * @param  [type]  $find_pattern [description]
-     * @param  [type]  $replacement  [description]
-     * @param  boolean $add          [description]
-     * @return [type]                [description]
+     *
+     * @param $find_pattern
+     * @param $replacement
+     * @param  boolean $add [description]
+     * @return void [type]                [description]
+     * @throws \Illuminate\Contracts\Filesystem\FileNotFoundException
      */
     protected function updateProviderBinding($find_pattern, $replacement, $add = true) {
         $service_provider = app_path('Providers/AppServiceProvider.php');
         $content = $this->filesystem->get($service_provider);
 
         $already_exists = preg_match($this->getBindPattern(), $content);
+
         if (($add && !$already_exists) || (!$add && $already_exists)) {
             $this->replaceFileContent($find_pattern, $replacement, $content, $service_provider);
         }
@@ -115,12 +117,12 @@ class MakePreparer extends BaseMakeCommand
 
     /**
      * Replace File content by prepared content
-     * 
-     * @param  [type] $find_pattern     [description]
-     * @param  [type] $replacement      [description]
-     * @param  [type] $content          [description]
-     * @param  [type] $service_provider [description]
-     * @return [type]                   [description]
+     *
+     * @param $find_pattern
+     * @param $replacement
+     * @param $content
+     * @param $service_provider
+     * @return void [type]                   [description]
      */
     protected function replaceFileContent($find_pattern, $replacement, $content, $service_provider)
     {
